monitoring/statsd-agent/requirements.txt
psutil==5.9.8
statsd-client==1.0.7
